Change some before_action to helper methods
Added method to current_user_check method in Application_helper that
determines whether processing can be performed vased on the authority of
the logged-in user

+++ app/controllers/categories_controller.rb
@@ -1,20 +1,21 @@
 class CategoriesController < ApplicationController
-  before_action :logged_in_user
-  before_action :set_categories, only:[:index]
   before_action :set_target_category, only:[:edit, :update, :destroy]
-  before_action :ensure_app_admin_or_admin, only:[:new, :create, :edit, :update, :destroy]
 
   def index
+    @categories = Category.all
   end
 
   def new
+    redirect_to categories_path unless current_user_check(1)
     @category = Category.new
   end
 
   def edit
+    redirect_to categories_path unless current_user_check(1)
   end
 
   def create
+    redirect_to categories_path unless current_user_check(1)
     @category = Category.new(category_params)
     if @category.save
       flash[:success] = "登録しました"
@@ -25,6 +26,7 @@ class CategoriesController < ApplicationController
   end
 
   def update
+    redirect_to categories_path unless current_user_check(1)
     if @category.update(category_params)
       flash[:success] = "更新しました"
       redirect_to categories_path
@@ -34,6 +36,7 @@ class CategoriesController < ApplicationController
   end
 
   def destroy
+    redirect_to categories_path unless current_user_check(1)
     if @category.destroy
       flash[:success] = "削除しました"
       redirect_to categories_path
@@ -43,14 +46,6 @@ class CategoriesController < ApplicationController
   end
 
   private
-  # アプリ管理者、管理者以外のユーザーをカテゴリー一覧ページにリダイレクト
-  def ensure_app_admin_or_admin
-    return if app_admin?(current_user)
-    return if admin?(current_user)
-    flash[:danger] = "権限がありません"
-    redirect_to categories_path
-  end
-
   def category_params
     params.require(:category).permit(:name)
   end
